+++ System Design/src/vehicleRentingSystemDesign/pricingStrategy/HourlyPricingStrategy.java
@@ -5,7 +5,7 @@ import vehicleRentingSystemDesign.vehicle.Vehicle;
 public class HourlyPricingStrategy implements PricingStrategy{
 
     @Override
-    public float calculatePricingStrategy(Vehicle vehicle, int hours){
+    public float calculatePricingStrategy(Vehicle vehicle, long hours){
         float pricePerDay = vehicle.getOneDayPrice();
         float priceOfHour = pricePerDay/(float)24.0;
