+++ System Design/src/vehicleRentingSystemDesign/pricingStrategy/PricingStrategy.java
@@ -3,5 +3,5 @@ package vehicleRentingSystemDesign.pricingStrategy;
 import vehicleRentingSystemDesign.vehicle.Vehicle;
 
 public interface PricingStrategy {
-    public float calculatePricingStrategy(Vehicle vehicle, int duration);
+    public float calculatePricingStrategy(Vehicle vehicle, long duration);
 }
